requirements.txt
fastapi
uvicorn

#for template
jinja2

#for static files
aiofiles

#for database   #new
sqlalchemy         
psycopg2

#for loading environment variables  #new
python-dotenv

#for email validation            #new
pydantic[email]

#logger
rich

#external
requests